feat: basic error handling feature for Mint dialog
