+++ lib/pages/scout/dialogs/athlete_page_dialogs.dart
@@ -1,2 +1,2 @@
-export 'mint_dialog.dart';
+export '../../../dialogs/mint/mint_dialog.dart';
 export 'redeem_dialog.dart';
