+++ lib/util/message.dart
@@ -1,0 +1,6 @@
+class Message {
+  static const String waitingApproval = 'Waiting for Approval';
+  static const String insufficient = 'Insufficient Balance';
+  static const String insufficientAX = 'Insufficient AX Balance';
+  static const String exception = 'Exception';
+}
